# Conflicts:
#	.github/workflows/ci-codeql-analysis.yml

+++ pi4j-core/src/main/java/com/pi4j/io/pwm/PwmPreset.java
@@ -1,37 +1,12 @@
 package com.pi4j.io.pwm;
 
-/*-
- * #%L
- * **********************************************************************
- * ORGANIZATION  :  Pi4J
- * PROJECT       :  Pi4J :: LIBRARY  :: Java Library (CORE)
- * FILENAME      :  PwmPreset.java
- *
- * This file is part of the Pi4J project. More information about
- * this project can be found here:  https://pi4j.com/
- * **********************************************************************
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-
 import com.pi4j.io.pwm.impl.DefaultPwmPresetBuilder;
 
 /**
- * <p>PwmPreset interface.</p>
- *
- * @author Robert Savage (<a href="http://www.savagehomeautomation.com">http://www.savagehomeautomation.com</a>)
- * @version $Id: $Id
+ * A named, reusable combination of PWM frequency and duty-cycle that can be applied
+ * to a {@link Pwm} instance on demand via {@link Pwm#applyPreset(String)}. Presets are
+ * built with a {@link PwmPresetBuilder} (see {@link #newBuilder(String)}) and may be
+ * registered through {@link PwmConfig} or {@link Pwm#addPreset(PwmPreset)}.
  */
 public interface PwmPreset {
 
@@ -62,18 +37,18 @@ public interface PwmPreset {
     }
 
     /**
-     *  Get the duty-cycle value as a decimal value that represents the
-     *  percentage of the ON vs OFF time of the PWM signal for each
-     *  period.  The duty-cycle range is valid from 0 to 100 including
-     *  factional values. (Values above 50% mean the signal will
-     *  remain HIGH more time than LOW.)
-     *
-     *  Example: A value of 50 represents a duty-cycle where half of
-     *  the time period the signal is LOW and the other half is HIGH.
+     * Get the duty-cycle value as a decimal value that represents the
+     * percentage of the ON vs OFF time of the PWM signal for each
+     * period.  The duty-cycle range is valid from 0 to 100 including
+     * factional values. (Values above 50% mean the signal will
+     * remain HIGH more time than LOW.)
+     * <p>
+     * Example: A value of 50 represents a duty-cycle where half of
+     * the time period the signal is LOW and the other half is HIGH.
      *
      * @return duty-cycle value expressed as a percentage (rage: 0-100)
      */
-    Float dutyCycle();
+    Integer dutyCycle();
 
     /**
      *  Get the duty-cycle value as a decimal value that represents the
@@ -81,13 +56,13 @@ public interface PwmPreset {
      *  period.  The duty-cycle range is valid from 0 to 100 including
      *  factional values.  (Values above 50% mean the signal will
      *  remain HIGH more time than LOW.)
-     *
+     * <p>
      *  Example: A value of 50 represents a duty-cycle where half of
      *  the time period the signal is LOW and the other half is HIGH.
      *
      * @return duty-cycle value expressed as a percentage (rage: 0-100)
      */
-    default Float getDutyCycle() {
+    default Integer getDutyCycle() {
         return dutyCycle();
     }
 
@@ -95,7 +70,7 @@ public interface PwmPreset {
      *  Get the configured frequency value in Hertz (number of cycles per second)
      *  that the PWM signal generator should attempt to output when this preset
      *  is applied to a PWM instance.
-     *
+     * <p>
      *  Please note that certain PWM signal generators may be limited to specific
      *  frequency bands and may not generate all possible explicit frequency values.
      *  After enabling the PWM signal using the 'on(...)' method, you can check the
@@ -111,7 +86,7 @@ public interface PwmPreset {
      *  Get the configured frequency value in Hertz (number of cycles per second)
      *  that the PWM signal generator should attempt to output when this preset
      *  is applied to a PWM instance.
-     *
+     * <p>
      *  Please note that certain PWM signal generators may be limited to specific
      *  frequency bands and may not generate all possible explicit frequency values.
      *  After enabling the PWM signal using the 'on(...)' method, you can check the
